better types setup
