+++ package.json
@@ -35,7 +35,7 @@
     "scripts": {
         "dev": "HOST=localhost PORT=3000 nuxt dev -c ./tests/fixtures/nuxt.config.js",
         "test": "ava",
-        "lint": "eslint lib/**/*.js types/**/*.ts",
+        "lint": "eslint --ext '.js,.ts,.vue' lib tests types",
         "release:patch": "sure && standard-version --no-verify --release-as patch && git push --tags",
         "release:minor": "sure && standard-version --no-verify --release-as minor && git push --tags",
         "release:major": "sure && standard-version --no-verify --release-as major && git push --tags",
