+++ tests/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "plugins": [
+        "ava",
+        "vue"
+    ]
+}
