+++ tests/api.flow.test.js
@@ -29,10 +29,10 @@ test('Test third level api (GET /api/v2/users/categories/types)', async (t) => {
 });
 
 test('Test hybrid api data flow server side.', async (t) => {
-    const window           = await nuxt.renderAndGetWindow(serverUrl('/'));
-    const path             = window.document.querySelector('.index span.path');
-    const okay             = window.document.querySelector('.index span.okay');
-    const resMiddle        = window.document.querySelector('.index span.response-middleware');
+    const window = await nuxt.renderAndGetWindow(serverUrl('/'));
+    const path = window.document.querySelector('.index span.path');
+    const okay = window.document.querySelector('.index span.okay');
+    const resMiddle = window.document.querySelector('.index span.response-middleware');
 
     t.is(okay.textContent, "It's okay!");
     t.is(resMiddle.textContent, "It's okay!");
@@ -49,16 +49,16 @@ test('Test hybrid api data flow server side.', async (t) => {
 });
 
 test('Test hybrid api data flow client side', async (t) => {
-    const window     = await nuxt.renderAndGetWindow(serverUrl('/'));
-    const path       = window.document.querySelector('.index span.path');
-    const okay       = window.document.querySelector('.index span.okay');
-    const idParam    = window.document.querySelector('.index span.id-param');
+    const window = await nuxt.renderAndGetWindow(serverUrl('/'));
+    const path = window.document.querySelector('.index span.path');
+    const okay = window.document.querySelector('.index span.okay');
+    const idParam = window.document.querySelector('.index span.id-param');
     const changePath = window.document.querySelector('.index .change-path');
-    const resMiddle  = window.document.querySelector('.index span.response-middleware');
+    const resMiddle = window.document.querySelector('.index span.response-middleware');
     const numOfUsers = window.document.querySelector('.index span.number-of-users');
     const createUser = window.document.querySelector('.index .create-user');
-    const firstUser  = window.document.querySelector('.index .first-user');
-    const ssError    = window.document.querySelector('.index .server-side-force-error');
+    const firstUser = window.document.querySelector('.index .first-user');
+    const ssError = window.document.querySelector('.index .server-side-force-error');
 
     changePath.dispatchEvent(new window.Event('click'));
     await new Promise(resolve => setTimeout(resolve, 2000)); // wait for API request
