+++ tests/fixtures/layouts/error.vue
@@ -4,6 +4,6 @@
 
 <script>
 export default {
-    props: ["error"]
+    props: ['error']
 };
 </script>
