--- tests/fixtures/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "baseUrl": ".",
-        "allowJs": true,
-        // "checkJs": true,
-        "allowSyntheticDefaultImports": true,
-        "target": "es2021",
-        "module": "ES2020",
-        "moduleResolution": "node",
-        "lib": [
-            "dom",
-            "es2021",
-        ],
-        "noEmit": true,
-        "strict": true,
-        "types": [
-            "@nuxt/types",
-        ],
-    },
-    "include": [
-        "**/*",
-        "../types/*.d.ts"
-    ]
-}
