+++ tsconfig.json
@@ -13,9 +13,16 @@
         ],
         "noEmit": true,
         "strict": true,
+        "types": [
+            "@nuxt/types",
+        ]
     },
+    "include": [
+        "./lib",
+        "./types/index.d.ts",
+        "./tests",
+    ],
     "exclude": [
-        "node_modules/**/*",
         "lib/plugins/api.template.js",
-    ]
+    ],
 }
