+++ types/index.d.ts
@@ -1,5 +1,3 @@
 import './module';
 import './extend';
 import './global';
-
-export * from './configuration';
